Separate Agents from Game Logic
In this commit, the original `ttt.c` was split to facilitate the
addition of new agents. Several changes were implemented to achieve
this:
1. `move_t` was introduced to transform `negamax` into a pure function
2. `qsort` was relocated from `available_moves` to incorporate it into
   the game logic.

The readability of the code also increases with the following
changes:
1. `table` and `move_record` were declared as static arrays.
2. The condition `is_consecutive(score, -1)` was replaced with
   `score < 0`, and vice versa.
